+++ packages/docs/src/stories/Button.stories.tsx
@@ -1,5 +1,5 @@
-import { Button, ButtonProps } from '@rcrdk-ignite-ui/react'
 import { ArrowRight } from '@phosphor-icons/react/dist/ssr'
+import { Button, ButtonProps } from '@rcrdk-ignite-ui/react'
 import type { Meta, StoryObj } from '@storybook/react'
 
 export default {