+++ packages/docs/src/stories/box.stories.tsx
@@ -1,11 +1,24 @@
-import { Box, BoxProps, Text } from '@rcrdk-ignite-ui/react'
+import { Box, BoxProps, Heading, Text } from '@rcrdk-ignite-ui/react'
 import type { Meta, StoryObj } from '@storybook/react'
 
 export default {
 	title: 'Surfaces/Box',
 	component: Box,
 	args: {
-		children: <Text>Testando o elemento box</Text>,
+		children: (
+			<>
+				<Heading css={{ marginBottom: '$4' }}>
+					Lorem ipsum dolor sit amet.
+				</Heading>
+
+				<Text>
+					Lorem ipsum dolor sit amet consectetur adipisicing elit. Earum
+					voluptatem excepturi possimus totam pariatur animi facere quos,
+					deserunt saepe quod dolorem est quasi sint, ratione modi, cupiditate
+					adipisci in nisi!
+				</Text>
+			</>
+		),
 	},
 	argTypes: {
 		children: {