+++ packages/docs/src/stories/heading.stories.tsx
@@ -7,10 +7,15 @@ export default {
 	args: {
 		children: 'Custom title',
 		size: 'md',
+		as: 'h2',
 	},
 	argTypes: {
 		size: {
 			options: ['sm', 'md', 'lg', '2xl', '4xl', '5xl', '6xl'],
+			control: 'inline-radio',
+		},
+		as: {
+			options: ['h1', 'h2', 'h3', 'h4', 'h5', 'h6', 'strong', 'em'],
 			control: 'inline-radio',
 		},
 	},