+++ packages/react/src/components/avatar/styles.ts
@@ -24,7 +24,7 @@ export const AvatarFallback = styled(Avatar.Fallback, {
 	width: '100%',
 	height: '100%',
 	background: '$gray600',
-	color: '$gray800',
+	color: '$gray400',
 
 	svg: {
 		width: '$6',