+++ packages/react/src/components/text.tsx
@@ -7,6 +7,8 @@ export const Text = styled('p', {
 	lineHeight: '$base',
 	margin: 0,
 	color: '$gray100',
+	boxSizing: 'border-box',
+	'-webkit-font-smoothing': 'antialiased',
 
 	variants: {
 		size: {