+++ packages/react/src/index.tsx
@@ -7,5 +7,7 @@ export * from './components/text-input'
 export * from './components/text-area'
 export * from './components/checkbox'
 export * from './components/multi-step'
+export * from './components/toast'
+export * from './components/tooltip'
 
 export * from './styles'