+++ packages/react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rcrdk-ignite-ui/react
 
+## 3.0.0
+
+### Major Changes
+
+- - Removed test token
+  - Add new tooltip component
+  - Add new toast component
+  - Improved components styles for better display
+
 ## 2.0.1
 
 ### Patch Changes