+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rcrdk-ignite-ui/tokens
 
+## 3.0.0
+
+### Major Changes
+
+- - Removed test token
+  - Add new tooltip component
+  - Add new toast component
+  - Improved components styles for better display
+
 ## 2.1.0
 
 ### Minor Changes